Backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
Authlib==1.5.2
autobahn==24.4.2
Automat==25.4.16
Bottleneck==1.5.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.3
daphne==4.2.0
dataclasses-json==0.6.7
Deprecated==1.2.18
Django==5.0.13
django-cors-headers==4.6.0
django-filter==24.3
django-jet-reboot==1.3.10
djangorestframework==3.15.2
docstring_parser==0.16
dotenv==0.9.9
fastapi==0.115.12
frozenlist==1.6.0
google-adk==0.4.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.92.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.14.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
graphviz==0.20.3
greenlet==3.2.1
grpc-google-iam-v1==0.14.2
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.6.1
incremental==24.7.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
langchain-google-vertexai==2.0.24
langchain-text-splitters==0.3.8
langsmith==0.3.42
Markdown==3.7
marshmallow==3.26.1
mcp==1.8.0
mssql-django==1.5
multidict==6.4.4
mypy_extensions==1.1.0
numexpr==2.10.2
numpy==2.2.5
oauth2client==4.1.3
opentelemetry-api==1.32.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
orjson==3.10.18
packaging==24.2
pillow==11.0.0
propcache==0.3.1
proto-plus==1.25.0
protobuf==5.29.2
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.27.2
pyodbc==5.2.0
pyOpenSSL==25.1.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
service-identity==24.2.0
setuptools==80.7.1
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.2
sse-starlette==2.3.4
starlette==0.46.2
tenacity==9.1.2
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.2
validators==0.35.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
zope.interface==7.2
zstandard==0.23.0
